PYIC-7026: Split nino contract tests
